+++ src/ResponsiveTable/ResponsiveTable.js
@@ -22,42 +22,42 @@ class ResponsiveTable extends Component {
 
     renderTableHeaders() {
         const {
-            headers,
+            columns,
         } = this.props;
 
-        if (!headers || !headers.length) {
+        if (!columns || !columns.length) {
             return null;
         }
 
         if (this.props.expandable) {
-            headers.push({ key: 'expandIcon', content: '' , alignRight: true });
+            columns.push({ key: 'expandIcon', header: '' , alignRight: true });
         }
 
-        return <TableHeaders headers={ headers } />;
+        return <TableHeaders columns={ columns } />;
     }
 
     renderTableFooter() {
         const {
-            headers,
-            headerData,
+            columns,
+            columnData,
         } = this.props;
 
-        if (!headers || !headers.length) {
+        if (!columns || !columns.length) {
           return null;
         }
 
-        if (headers.findIndex((header) => 'footerContent' in header) === -1 ) {
+        if (!columns.some(column => column.footer)) {
           return null;
         }
 
-        return <TableFooter headers={ headerData || headers } />;
+        return <TableFooter columns={ columnData || columns } />;
     }
 
     renderTableBody() {
         const {
             data,
-            headers,
-            headerData
+            columns,
+            columnData
         } = this.props;
 
         if (this.props.children) {
@@ -71,7 +71,7 @@ class ResponsiveTable extends Component {
         return (
             <tbody>
                 {data.map((row, index) => (
-                    <TableRow cells={ row } headers={ headerData || headers } key={ index } />
+                    <TableRow cells={ row } columns={ columnData || columns } key={ index } />
                 ))}
             </tbody>
         );
@@ -93,15 +93,15 @@ ResponsiveTable.propTypes = {
     caption: PropTypes.string,
     expandable: PropTypes.bool,
     children: PropTypes.node,
-    headerData: PropTypes.arrayOf(
+    columnData: PropTypes.arrayOf(
         PropTypes.object
     ),
     data: PropTypes.arrayOf(
             PropTypes.object
         ),
-    headers: PropTypes.arrayOf(
+    columns: PropTypes.arrayOf(
         PropTypes.shape({
-            content: PropTypes.node.isRequired,
+            header: PropTypes.node.isRequired,
             key: PropTypes.string.isRequired,
         }).isRequired
     ),