+++ CHANGELOG.md
@@ -1,9 +1,32 @@
 # Changelog
 
-## v1.0.3
+## v2.0.0
 
-* table footers
+* BREAKING: `headers` prop is renamed to `columns` to include footers
 * data is no longer a required prop in ResponsiveTable
+
+## Migrating to v.2.0.0
+* Rename `headers` prop to `columns` in all tables
+* Rename the `content` prop in the columns objects to `header`
+
+change this:
+
+```javascript
+const headers = [
+    { key: 'name', content: 'Varenavn' },
+    { key: 'price', content: 'Pris', alignRight: true }
+];
+<ResponsiveTable headers={headers} data={data}>
+```
+to this:
+```javascript
+const columns = [
+    { key: 'name', header: 'Varenavn' },
+    { key: 'price', header: 'Pris', alignRight: true }
+];
+<ResponsiveTable columns={columns} data={data}>
+
+```
 
 ## v1.0.2